requirements.txt
Django==1.8
psycopg2==2.6
djangorestframework==3.0
drf-extensions==0.2.4